refactor: DTO를 common 모듈로 이전
