+++ common-module/src/main/java/hongik/triple/commonmodule/dto/analysis/YoutubeVideoDto.java
@@ -1,0 +1,9 @@
+package hongik.triple.commonmodule.dto.analysis;
+
+public record YoutubeVideoDto(
+        String videoId,
+        String videoTitle,
+        String videoUrl,
+        String channelName,
+        String thumbnailUrl
+) {}
